Move URL related functions into URL utils
